refactor(product-list): reorganize file structure and fix useSearchFilter to navigate within the same user locale
