--- src/app/[locale]/(public)/products/_components/pagination-wrapper.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import Pagination from '@/components/shared/pagination';
-
-export default function PaginationWrapper({
-  totalPages,
-}: {
-  totalPages: number;
-}) {
-  return <Pagination totalPages={totalPages} />;
-}
