+++ src/i18n/messages/ar.json
@@ -7,7 +7,6 @@
   "login-button": "تسجيل الدخول",
   "register-paragraph": "هل ليس لديك اي حساب ؟",
   "register-paragraph-action": "قم بعمل حساب الان !",
-  
   "forgot-password-title": "نسيت كلمة المرور",
   "forgot-password-description": "لا تقلق، سنرسل لك تعليمات لإعادة تعيين كلمة المرور.",
   "forgot-password-redirect-message": "هل لديك حساب بالفعل؟",
@@ -41,9 +40,9 @@
   "passwords-mismatch": "كلمتا المرور غير متطابقتين",
   "password-confirm-required": "يرجى تأكيد كلمة المرور",
   "send-otp-button": "إرسال رمز جديد",
-  "send-otp-button-time-count": "أرسل رمزًا جديدًا خلال: {{timeLeft}} ثانية",
+  "send-otp-button-time-count": "أرسل رمزًا جديدًا خلال: {timeLeft} ثانية",
   "enter-otp-label": "أدخل رمز OTP",
-  "otp-description": "تم إرسال رمز مكون من ٦ أرقام إلى {email} <edit>تعديل</edit>",
+  "otp-edit-email": "تم إرسال رمز مكون من ٦ أرقام إلى {email} <edit>تعديل</edit>",
   "need-help": "بحاجة الى مساعدة؟",
   "contact-us-label": "اتصل بنا",
   "verify-otp": "التحقق من كلمة المرور لمرة واحدة (OTP).",
