+++ src/lib/actions/otp-verification-action.ts
@@ -2,7 +2,7 @@
 
 import { REQUEST_HEADERS } from '../constants/request-headers.constant';
 
-//  Mark single notification as read
+//  Sent otp to verify
 async function otpVerification(resetCode: string) {
   const API_URL = `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/v1/auth/verifyResetCode`;
